Move legacy card types and interfaces to legacy/ directory in libs/auth
