+++ libs/auth/package.json
@@ -35,6 +35,7 @@
     "@votingworks/logging": "workspace:*",
     "@votingworks/types": "workspace:*",
     "@votingworks/utils": "workspace:*",
+    "base64-js": "^1.3.1",
     "node-fetch": "^2.6.0",
     "zod": "3.14.4"
   },
@@ -50,6 +51,7 @@
     "eslint-plugin-jest": "^26.1.5",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-vx": "workspace:*",
+    "fetch-mock": "^9.11.0",
     "is-ci-cli": "^2.2.0",
     "jest": "^27.3.1",
     "jest-junit": "^14.0.1",
