+++ libs/utils/src/index.ts
@@ -2,7 +2,6 @@
 export * from './as_boolean';
 export * from './auth';
 export * from './ballot_package';
-export * from './Card';
 export * from './pins';
 export * from './compressed_tallies';
 export * from './date';
